requirements.txt
anyio==4.13.0
certifi==2026.2.25
charset-normalizer==3.4.7
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-telegram-bot==22.7
requests==2.33.1
urllib3==2.6.3
